completed css and basic JS implemented

+++ App.js
@@ -4,13 +4,28 @@ function authenticateUserRegistration(ev){
     var userN = document.getElementById("loginId").value;
     var passW = document.getElementById("currentp").value;
     localStorageOfCredentials(userN, passW);
-    alert("Successfully Registered...");
-
 }
 
 function localStorageOfCredentials(email, password){
-    localStorage.setItem(email,password);
+    if(!userExist(email)){
+        localStorage.setItem(email,password);
+        alert("Successfully Registered...");
+        location.replace("https://formafitnessclub.netlify.app/");
+    }else{
+        document.getElementById("loginId").value = "";
+        document.getElementById("currentp").value = "";
+    }
 }  // stored email and password within local storage
+
+function userExist(email){
+    try{
+        let x = localStorage.getItem(email);
+        alert(`${email} already registered, use different email ID.`);
+        return true;
+    } catch{
+        return false;
+    }
+}
 
 function authenticateUserLogin(e){
     e.preventDefault();
@@ -27,6 +42,7 @@ function authenticateUserLogin(e){
         alert("Authorized");
         document.getElementById("loginId").value = "";
         document.getElementById("currentp").value = "";
+        location.replace("https://formafitnessclub.netlify.app/");
     } else{
         alert("invalid credentials/ password");
     }
